BNO055/requirements.txt
numpy
pandas
matplotlib
scikit-learn
jupyter
adafruit-circuitpython-bno055
pyserial
quaternion
